Project
* Clean

+++ Engine/Framework/Engine/ECS/Components/Transform.cs
@@ -10,6 +10,6 @@ namespace Hybrid
 
         public Vector2 Position = Vector2.Zero;
         public Vector2 Scale = Vector2.One;
-        public int Rotation = 0;
+        public float Rotation = 0;
     }
 }
